Fix all errors and bugs
