+++ Gemfile
@@ -9,6 +9,8 @@ gem "rails", "~> 7.0.4", ">= 7.0.4.2"
 # The original asset pipeline for Rails [https://github.com/rails/sprockets-rails]
 gem 'bootstrap', '~> 5.2.2'
 
+gem 'draper'
+
 gem "sprockets-rails"
 
 # Use sqlite3 as the database for Active Record
