+++ app/controllers/questions_controller.rb
@@ -20,8 +20,10 @@ class QuestionsController < ApplicationController
     end
 
     def show 
+      @question = @question.decorate
       @answer = @question.answers.build
       @pagy, @answers =  pagy @question.answers.order(created_at: :desc)
+      @answers = @answers.decorate
     end 
 
     def edit
@@ -40,6 +42,7 @@ class QuestionsController < ApplicationController
 
     def index
       @pagy, @questions = pagy Question.order(created_at: :desc)
+      @questions = @questions.decorate
     end 
 
     def new
