+++ app/controllers/users_controller.rb
@@ -7,7 +7,7 @@ class UsersController < ApplicationController
       @user = User.new user_params
       if @user.save
         session[:user_id] = @user.id
-        flash[:success] = "Welcome to the app, #{@user.name}!"
+        flash[:success] = "Welcome to the app, #{current_user.name_or_email}!"
         redirect_to root_path
       else
         render :new
