+++ app/decorators/answer_decorator.rb
@@ -1,0 +1,8 @@
+class AnswerDecorator < ApplicationDecorator
+  delegate_all
+  
+  def formatted_created_at
+    created_at.strftime('%Y-%m-%d %H:%M')
+  end
+
+end
