+++ app/decorators/application_decorator.rb
@@ -1,3 +1,3 @@
 class ApplicationDecorator < Draper::Decorator
-  
+    
 end
