+++ app/models/answer.rb
@@ -3,7 +3,4 @@ class Answer < ApplicationRecord
 
   validates :body, presence: true
 
-  def formatted_created_at
-    created_at.strftime('%Y-%m-%d %H:%M')
- end
 end
