+++ app/models/user.rb
@@ -2,6 +2,5 @@ class User < ApplicationRecord
     has_secure_password
     
     validates :email, presence: true, uniqueness: true
-    validates :name, presence: true 
 
 end
