+++ app/views/layouts/application.html.erb
@@ -23,12 +23,11 @@
           <% if user_signed_in? %>
             <li>
               <a class="nav-link px-2 text-white" href="#">
-                <%= current_user.name %>
+                <%= current_user.name_or_email %>
               </a>
             </li>
           <% else %>
-            <li><%= nav_tab 'Sign Up', new_user_path,
-              class: 'nav-link px-2', current_page: current_page %></li>
+            <li><%= link_to 'Sign Up', new_user_path, class: 'btn btn-light' %></li>
           <% end %>
         </ul>
         </nav>
